+++ packages/kit/types/internal.d.ts
@@ -10,9 +10,18 @@ import {
 	HandleError,
 	RequestEvent,
 	RequestOptions,
-	PrerenderErrorHandler,
 	Server
 } from './index';
+import {
+	Either,
+	Fallthrough,
+	HttpMethod,
+	JSONObject,
+	MaybePromise,
+	ResponseHeaders,
+	RouteSegment,
+	TrailingSlash
+} from './private';
 
 export interface ServerModule {
 	Server: typeof InternalServer;
@@ -33,8 +42,6 @@ export interface Asset {
 	size: number;
 	type: string | null;
 }
-
-export type Body = JSONValue | Uint8Array | ReadableStream | import('stream').Readable;
 
 export interface BuildData {
 	app_dir: string;
@@ -65,8 +72,6 @@ export type CSRComponentLoader = () => Promise<CSRComponent>;
 
 export type CSRRoute = [RegExp, CSRComponentLoader[], CSRComponentLoader[], GetParams?, HasShadow?];
 
-export type Either<T, U> = Only<T, U> | Only<U, T>;
-
 export interface EndpointData {
 	type: 'endpoint';
 	key: string;
@@ -74,10 +79,6 @@ export interface EndpointData {
 	pattern: RegExp;
 	params: string[];
 	file: string;
-}
-
-export interface Fallthrough {
-	fallthrough: true;
 }
 
 export type GetParams = (match: RegExpExecArray) => Record<string, string>;
@@ -91,8 +92,6 @@ export interface Hooks {
 	handleError: HandleError;
 }
 
-export type HttpMethod = 'get' | 'head' | 'post' | 'put' | 'delete' | 'patch';
-
 export class InternalServer extends Server {
 	respond(
 		request: Request,
@@ -102,19 +101,6 @@ export class InternalServer extends Server {
 	): Promise<Response>;
 }
 
-export type JSONObject = { [key: string]: JSONValue };
-
-export type JSONValue = string | number | boolean | null | ToJSON | JSONValue[] | JSONObject;
-
-export interface Logger {
-	(msg: string): void;
-	success(msg: string): void;
-	error(msg: string): void;
-	warn(msg: string): void;
-	minor(msg: string): void;
-	info(msg: string): void;
-}
-
 export interface ManifestData {
 	assets: Asset[];
 	layout: string;
@@ -122,8 +108,6 @@ export interface ManifestData {
 	components: string[];
 	routes: RouteData[];
 }
-
-export type MaybePromise<T> = T | Promise<T>;
 
 export interface MethodOverride {
 	parameter: string;
@@ -142,8 +126,6 @@ export type NormalizedLoadOutput = Either<
 	Fallthrough
 >;
 
-type Only<T, U> = { [P in keyof T]: T[P] } & { [P in Exclude<keyof U, keyof T>]?: never };
-
 export interface PageData {
 	type: 'page';
 	key: string;
@@ -161,51 +143,17 @@ export interface PrerenderDependency {
 	body: null | string | Uint8Array;
 }
 
-export type PrerenderOnErrorValue = 'fail' | 'continue' | PrerenderErrorHandler;
-
 export interface PrerenderOptions {
 	fallback?: string;
 	all: boolean;
 	dependencies: Map<string, PrerenderDependency>;
 }
 
-export type RecursiveRequired<T> = {
-	// Recursive implementation of TypeScript's Required utility type.
-	// Will recursively continue until it reaches primitive or union
-	// with a Function in it, except those commented below
-	[K in keyof T]-?: Extract<T[K], Function> extends never // If it does not have a Function type
-		? RecursiveRequired<T[K]> // recursively continue through.
-		: K extends 'vite' // If it reaches the 'vite' key
-		? Extract<T[K], Function> // only take the Function type.
-		: T[K]; // Use the exact type for everything else
-};
-
-export interface RequiredResolveOptions {
-	ssr: boolean;
-	transformPage: ({ html }: { html: string }) => MaybePromise<string>;
-}
-
 export interface Respond {
 	(request: Request, options: SSROptions, state?: SSRState): Promise<Response>;
 }
 
-/** `string[]` is only for set-cookie, everything else must be type of `string` */
-export type ResponseHeaders = Record<string, string | number | string[]>;
-
 export type RouteData = PageData | EndpointData;
-
-export interface RouteDefinition {
-	type: 'page' | 'endpoint';
-	pattern: RegExp;
-	segments: RouteSegment[];
-	methods: HttpMethod[];
-}
-
-export interface RouteSegment {
-	content: string;
-	dynamic: boolean;
-	rest: boolean;
-}
 
 export interface ShadowEndpointOutput<Output extends JSONObject = JSONObject> {
 	status?: number;
@@ -341,8 +289,5 @@ export interface SSRState {
 
 export type StrictBody = string | Uint8Array;
 
-type ToJSON = { toJSON(...args: any[]): Exclude<JSONValue, ToJSON> };
-
-export type TrailingSlash = 'never' | 'always' | 'ignore';
-
 export * from './index';
+export * from './private';
